+++ packages/react-csrf-universal-provider/src/CsrfUniversalProvider.ts
@@ -1,6 +1,7 @@
-import {CsrfTokenContext} from '@shopify/react-csrf';
+import {CsrfTokenContext, CsrfTokenData} from '@shopify/react-csrf';
 import {createUniversalProvider} from '@shopify/react-universal-provider';
 
-export const CsrfUniversalProvider = createUniversalProvider<
-  string | undefined
->('csrf-token', CsrfTokenContext);
+export const CsrfUniversalProvider = createUniversalProvider<CsrfTokenData>(
+  'csrf-token',
+  CsrfTokenContext,
+);