+++ packages/react-csrf/src/hooks.ts
@@ -5,9 +5,9 @@ import {CsrfTokenContext} from './context';
 export function useCsrfToken() {
   const csrf = useContext(CsrfTokenContext);
 
-  if (csrf == null) {
+  if (!csrf || csrf.csrfToken === undefined) {
     throw new Error('No CSRF token found in context.');
   }
 
-  return csrf;
+  return csrf.csrfToken;
 }