+++ packages/react-csrf/CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+- Changed context type to take { csrfToken: string | undefined; } to work properly with newest serialize-javascript version 3.0.0 and JSON.parse enforcement
 
 ## [1.0.8] - 2020-10-15